+++ src/main/java/com/gofobao/framework/asset/entity/YesterdayAsset.java
@@ -19,19 +19,19 @@ public class YesterdayAsset {
     private Long userId;
     @Basic
     @Column(name = "use_money")
-    private Integer useMoney;
+    private Long useMoney;
     @Basic
     @Column(name = "no_use_money")
-    private Integer noUseMoney;
+    private Long noUseMoney;
     @Basic
     @Column(name = "virtual_money")
-    private Integer virtualMoney;
+    private Long virtualMoney;
     @Basic
     @Column(name = "collection")
-    private Integer collection;
+    private Long collection;
     @Basic
     @Column(name = "payment")
-    private Integer payment;
+    private Long payment;
     @Basic
     @Column(name = "updated_at")
     private Date updatedAt;
