+++ src/main/java/com/gofobao/framework/asset/vo/response/VoCollectionResp.java
@@ -14,15 +14,15 @@ public class VoCollectionResp extends VoBaseResp{
     @ApiModelProperty("待收利息")
     private String interest;
     @ApiModelProperty("不显示待收利息")
-    private Integer hideInterest;
+    private Long hideInterest;
 
     @ApiModelProperty("待收本金")
     private String principal;
     @ApiModelProperty("不显示待收本金")
-    private Integer hidePrincipal;
+    private Long hidePrincipal;
 
     @ApiModelProperty("待收总金额")
     private String waitCollectionTotal;
     @ApiModelProperty("不显示待收总金额")
-    private Integer hideWaitCollectionTotal;
+    private Long hideWaitCollectionTotal;
 }
