+++ src/main/java/com/gofobao/framework/integral/biz/impl/IntegralBizImpl.java
@@ -118,7 +118,7 @@ public class IntegralBizImpl implements IntegralBiz {
             e.printStackTrace();
         }
 
-        Integer collection = asset.getCollection();//代收金额
+        Long collection = asset.getCollection();//代收金额
         Integer totalIntegral = integral.getUseIntegral() + integral.getNoUseIntegral();//总积分
 
         VoListIntegralResp voListIntegralResp = VoBaseResp.ok("查询成功", VoListIntegralResp.class);
@@ -211,7 +211,7 @@ public class IntegralBizImpl implements IntegralBiz {
             e.printStackTrace();
         }
 
-        Integer collection = asset.getCollection();
+        Long collection = asset.getCollection();
         Integer sumIntegral=useIntegral+integral.getNoUseIntegral();
         String takeRatesStr = getTakeRates(collection, sumIntegral, integralRule);
         double takeRates = Double.parseDouble(takeRatesStr);//折现系数
@@ -288,7 +288,7 @@ public class IntegralBizImpl implements IntegralBiz {
      * @param integral 总积分
      * @return
      */
-    private String getTakeRates(Integer money, Integer integral, List<Map<String, String>> maps) {
+    private String getTakeRates(Long money, Integer integral, List<Map<String, String>> maps) {
         money = money / 100;
 
         Integer moneyMin = 0;
